Update summary_generator.py
Made updates to break apart the filtering of topics (ex. aerospace) from the filtering of stories (ex. Apple releases a new iPhone).

Also added a secondary dedupe. It works the same as the original with a slightly different prompt that is meant to catch items that might slip through the first dedupe.